have LiveDataRegistry register an invalidation callback for itself
