+++ gto-support-db-livedata/src/main/kotlin/org/ccci/gto/android/common/db/LiveDataDao.kt
@@ -2,17 +2,14 @@ package org.ccci.gto.android.common.db
 
 import android.annotation.SuppressLint
 import android.database.Cursor
-import androidx.annotation.AnyThread
 import androidx.annotation.MainThread
 import androidx.annotation.RestrictTo
-import androidx.collection.SimpleArrayMap
 import androidx.lifecycle.ComputableLiveData
 import androidx.lifecycle.LiveData
-import java.util.WeakHashMap
 
 interface LiveDataDao : Dao {
     @get:RestrictTo(RestrictTo.Scope.SUBCLASSES)
-    val liveDataRegistry get() = getService { LiveDataRegistry() }
+    val liveDataRegistry get() = getService { LiveDataRegistry(this) }
 
     @MainThread
     @SuppressLint("RestrictedApi")
@@ -55,31 +52,6 @@ private class DaoGetCursorComputableLiveData<T : Any>(dao: LiveDataDao, private 
     override fun compute() = dao.getCursor(query)
 }
 // endregion DaoComputableLiveData
-
-class LiveDataRegistry {
-    private val registry: SimpleArrayMap<Class<*>, MutableMap<ComputableLiveData<*>, Unit>> = SimpleArrayMap()
-
-    @MainThread
-    internal fun DaoComputableLiveData<*>.registerFor(clazz: Class<*>) {
-        synchronized(registry) {
-            (registry[clazz] ?: WeakHashMap<ComputableLiveData<*>, Unit>().also { registry.put(clazz, it) })[this] =
-                Unit
-        }
-    }
-
-    @MainThread
-    internal fun DaoComputableLiveData<*>.registerFor(query: Query<*>) {
-        query.allTables.forEach { registerFor(it.type) }
-    }
-
-    @AnyThread
-    @SuppressLint("RestrictedApi")
-    fun invalidate(clazz: Class<*>) {
-        synchronized(registry) {
-            registry[clazz]?.keys?.forEach { it.invalidate() }
-        }
-    }
-}
 
 inline fun <reified T : Any> LiveDataDao.findLiveData(vararg key: Any) = findLiveData(T::class.java, *key)
 fun <T : Any> Query<T>.getAsLiveData(dao: LiveDataDao) = dao.getLiveData(this)
